Move RawFileTransmissionData to schema
