+++ transmission/transmission_plugin/uv_vis_nir_transmission_plugin/tests/data/stable_version/RawFileTransmission.archive.yaml
@@ -1,0 +1,2 @@
+data:
+  m_def: uv_vis_nir_transmission.parser.RawFileTransmissionData
